create conftest add fixture for generators
